# core/docs/swagger.yaml
definitions:
...
  internal_api_handlers.VlanCreateRequest:
    properties:
      id:
        type: integer
      ip:
        type: string
      management:
        items:
          type: string
        type: array
      name:
        type: string
      parent:
        type: string
      state:
        type: string
      zone:
        type: string
    type: object
  internal_api_handlers.VlanEditRequest:
    properties:
      id:
        type: integer
      ip:
        type: string
      management:
        items:
          type: string
        type: array
      parent:
        type: string
      state:
        type: string
      zone:
        type: string
    type: object
  internal_api_handlers.ZoneCreateRequest:
    properties:
      color:
        type: string
      interfaces:
        items:
          type: string
        type: array
      name:
        type: string
      networks:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  internal_api_handlers.ZoneEditRequest:
    properties:
      color:
        type: string
      interfaces:
        items:
          type: string
        type: array
      networks:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
...
  sentinelos_core_internal_model.Vlan:
    properties:
      id:
        type: integer
      ip:
        type: string
      management:
        items:
          type: string
        type: array
      name:
        type: string
      parent:
        type: string
      state:
        type: string
      zone:
        type: string
    type: object
  sentinelos_core_internal_model.Zone:
    properties:
      color:
        type: string
      interfaces:
        items:
          type: string
        type: array
      name:
        type: string
      networks:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/sentinelos_core_internal_model.ZoneType'
    type: object
  sentinelos_core_internal_model.ZoneType:
    enum:
    - l2
    - l3
    type: string
    x-enum-varnames:
    - ZoneL2
    - ZoneL3
...
  sentinelos_core_pkg_utils.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Next-Generation Firewall Management API.
  title: SentinelOS API
  version: "1.0"
paths:
...
  /api/vlans:
    get:
      description: Get a list of all vlans.
      produces:
      - application/json
      responses:
        "200":
          description: List of vlans
          schema:
            additionalProperties:
              $ref: '#/definitions/sentinelos_core_internal_model.Vlan'
            type: object
        "500":
          description: ERR_SYS_4001 Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: List Vlans
      tags:
      - networking
      - vlans
    post:
      consumes:
      - application/json
      description: Create a new vlan object.
      parameters:
      - description: Vlan details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.VlanCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: vlan created in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Zone does not exist
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create Vlan
      tags:
      - networking
      - vlans
  /api/vlans/{name}:
    delete:
      description: Delete an existing vlan object.
      parameters:
      - description: Vlan Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: vlan deleted from candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete Vlan
      tags:
      - networking
      - vlans
    put:
      consumes:
      - application/json
      description: Update an existing vlan object.
      parameters:
      - description: Vlan Name
        in: path
        name: name
        required: true
        type: string
      - description: Vlan details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.VlanEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: vlan updated in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Zone does not exist
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Edit Vlan
      tags:
      - networking
      - vlans
  /api/zones:
    get:
      description: Get a list of all zones.
      produces:
      - application/json
      responses:
        "200":
          description: List of zones
          schema:
            additionalProperties:
              $ref: '#/definitions/sentinelos_core_internal_model.Zone'
            type: object
        "500":
          description: ERR_SYS_4001 Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: List Zones
      tags:
      - security
      - zones
    post:
      consumes:
      - application/json
      description: Create a new zone object.
      parameters:
      - description: Zone details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.ZoneCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: zone added in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Zone Type
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "409":
          description: ERR_NET_2006 Resource already exists
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create Zone
      tags:
      - security
      - zones
  /api/zones/{name}:
    delete:
      description: Delete an existing zone object.
      parameters:
      - description: Zone Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: zone deleted in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete Zone
      tags:
      - security
      - zones
    put:
      consumes:
      - application/json
      description: Update an existing zone object.
      parameters:
      - description: Zone Name
        in: path
        name: name
        required: true
        type: string
      - description: Zone details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.ZoneEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: zone updated in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Zone Type
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Edit Zone
      tags:
      - security
      - zones
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"